pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xyz-agent-context"
version = "1.3.1"
description = "XYZ Agent Context Framework for multi-agent systems"
readme = "README.md"
license = {text = "CC-BY-NC-4.0"}
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.72.0",
    "claude-agent-sdk>=0.1.6",
    "openai>=2.7.1",
    "openai-agents>=0.5.0",
    "mcp[cli]>=1.20.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.2.0",
    "croniter>=6.0.0",
    "aiomysql>=0.3.2",
    "aiosqlite>=0.20.0",
    "google-genai>=1.0.0",
    "fastmcp>=2.14.1",
    "numpy>=1.26.0",
    "httpx[socks]",
    "bcrypt>=5.0.0",
    "pyjwt>=2.10.1",
    "lark-oapi>=1.4.0,<2.0.0",
    "ddgs>=9.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/xyz_agent_context", "backend"]

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "pillow>=12.2.0",
    "psutil>=7.2.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
# Start conservative — only catch clear bugs, not style
select = ["E", "F"]
# Ignore rules that conflict with existing codebase patterns
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module-level import not at top (lazy imports are intentional)
    "E721",   # type comparison using == (existing patterns)
    "E741",   # ambiguous variable name
    "F401",   # unused imports (too many false positives with re-exports)
    "F811",   # redefinition of unused name
    "F821",   # undefined name (false positives with forward refs and method scope)
    "F841",   # local variable assigned but never used
]

[tool.ruff.lint.isort]
known-first-party = ["xyz_agent_context", "backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
